Fix size_t to int cast warnings.
Several instances in Visual Studio 2015:

warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

+++ glslang/MachineIndependent/Intermediate.cpp
@@ -2241,7 +2241,7 @@ bool TIntermediate::promoteAggregate(TIntermAggregate& node)
 {
     TOperator op = node.getOp();
     TIntermSequence& args = node.getSequence();
-    const int numArgs = args.size();
+    const int numArgs = static_cast<int>(args.size());
 
     // Presently, only hlsl does intrinsic promotions.
     if (getSource() != EShSourceHlsl)
